prevention-engine/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
scapy==2.5.0
python-dotenv==1.0.1
requests==2.31.0
pyyaml==6.0.1
reportlab==4.1.0
dpkt==1.9.8
